move `ACTION_LINK` into `constants` lib file
